# src/resources/Swagger-JSON-Reference.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "PORTAL-BCA",
    "description": "",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/profiles/updateprofile/:profileid": {
      "put": {
        "description": "",
        "summary": "UPDATE PROFILE",
        "tags": [
          "PROFILES"
        ],
        "operationId": "UPDATEPROFILE",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UPDATEPROFILERequest"
            }
          },
          {
            "name": "Content-Type", "default":"application/json",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            },
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UPDATEPROFILERequest": {
      "title": "UPDATEPROFILERequest",
      "example": {
        "title": "Revisões"
      },
      "type": "object"
    },
...
  },
...
}